+++ cluster_agent/identity/slurm_user/mappers/__init__.py
@@ -1,0 +1,10 @@
+from cluster_agent.identity.slurm_user.mappers.mapper_base import SlurmUserMapper
+from cluster_agent.identity.slurm_user.mappers.ldap import LDAPMapper
+from cluster_agent.identity.slurm_user.mappers.single_user import SingleUserMapper
+
+
+__all__ = [
+    "SlurmUserMapper",
+    "LDAPMapper",
+    "SingleUserMapper",
+]