+++ cluster_agent/utils/exception.py
@@ -2,29 +2,25 @@
 from buzz import Buzz
 
 
-class ProcessExecutionError(Buzz):
+class ClusterAgentError(Buzz):
     """Raise exception when execution command returns an error"""
 
 
-class AuthTokenError(Buzz):
+class ProcessExecutionError(ClusterAgentError):
+    """Raise exception when execution command returns an error"""
+
+
+class AuthTokenError(ClusterAgentError):
     """Raise exception when there are connection issues with the backend"""
 
 
-class SlurmrestdError(Buzz):
+class SlurmrestdError(ClusterAgentError):
     """Raise exception when slurmrestd raises any error"""
 
 
-class JobbergateApiError(Buzz):
+class JobbergateApiError(ClusterAgentError):
     """Raise exception when communication with Jobbergate API fails"""
 
 
-class JobSubmissionError(Buzz):
+class JobSubmissionError(ClusterAgentError):
     """Raise exception when a job cannot be submitted raises any error"""
-
-
-class LDAPError(Buzz):
-    """Raise exception when LDAP communication fails."""
-
-
-class UIDError(Buzz):
-    """Raise exception when a UID (and GID) cannot be fetched."""